Fix: valor percentual em decision

+++ asgard/workers/models/decision.py
@@ -1,5 +1,5 @@
 class Decision:
-    def __init__(self, appid: str, cpu: float = None, mem: int = None):
+    def __init__(self, appid: str, cpu: float = None, mem: float = None):
         self.id = appid
         self.cpu = cpu
         self.mem = mem
